Implemented insert function
insert: Will take a value and take it to the most appropriate leaf.
First, it will check if it equals the root node, if it does it will
return. If it is less than that node, we will then iteratively go
through the same process on the left or the right if it's greater than
the root node.

 Changes to be committed:
	modified:   binarySearchTree.js