Implemented height method
height: for a given node (root if non provided), the function will check
if there is either a node to the left or right. Increment the respective
half that is available and then recursilvey call itself on that node. If
there is no node to recurse, the function will return the max of the
left or eight height.

 Changes to be committed:
	modified:   binarySearchTree.js