Implemented isBalanced and rebalance methods
First, updated the root to return in a getter function to enable
rebalance to return the updated root.

isBalanced: checks if the tree is balanced by checking if the heights of
the left and right subtrees differ by at most 1 or not.

rebalance: uses the preorder traversal to return an array of the tree
elements. We then go ahead and sort and dedupe that array before
updating the root node with a new buildTree.

 Changes to be committed:
	modified:   binarySearchTree.js